Add HTMLCanvasElement as accepted type for image and move path stuff back
