+++ nusight2/src/client/components/camera/objects/text/README.md
