+++ nusight2/src/client/components/camera/objects/text/shaders/text.vert
