+++ nusight2/src/client/components/camera/objects/text/stories/text_interactive.stories.tsx
